[355310] Add border to 'Use' Combo on extension point schema attribute page.
This bug fix is in two parts - first, make sure that the toolkit is
painting the borders for the two sub-composites in the 'Use' section.

Second, modify the borders of those composites to add a little more room
around them, because this is where the border is painted - see the
Javadoc of
org.eclipse.ui.forms.widgets.FormToolkit.paintBordersFor(Composite),
which says "Care should be taken when selection layout margins. At least
one pixel margin width and height must be chosen to allow the toolkit to
paint the border on the parent around the widgets." In the main, most of
the elements in this form don't have any additional margin but it
doesn't matter because they're not closely packed in a composite. Note
that this solution makes the right-hand end of the Use combo one pixel
short of lining up with the rest of the elements. However, this is less
destructive than changing all the other elements' helper methods, etc.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=355310
Change-Id: I7913a10b5c585a3c48cd393f6f95feb2c0e2adc8
Signed-off-by: Ben Cox <benjamin.cox@uk.ibm.com>